Modified ctlrequest/raftprocess lookup to use peer UUID.
As per the discussion, recipes will pass the peer UUID
directly to perform ctlrequest/raftprocess operations.

TODO: Still needs to add the code where we will get
the non-running peer UUID rather than accessing the
peer-array which we prepare at the start of the recipe.
Upload that change in next commit.