cmd and where should be passed as cmdline argument.
Tried to access the cmd and where from the ansible cache
in lookup call. But if we want to create the cmd on the
fly by preparing string with some varibles. Using cache
variable is not possible. Because the string value is not
evaluated before accessing through lookup plugin.
So passing the cmd and where values from cmdline.