Recipes should pass APPLY and WHERE cmd text. (#67)
* Recipes should pass APPLY and WHERE cmd text.

- Prevously we had ctlrequest dictionary for the
  command string.
- But with more number of APPLY cmds, it makes more
  sense for recipes to pass exact text for APPLY cmd
  and for WHERE.
- Added common task for start_server which will loop
  till the bootup completes for the server.
- Some corrections with spaces in the range cmd.

* cmd and where should be passed as cmdline argument.

Tried to access the cmd and where from the ansible cache
in lookup call. But if we want to create the cmd on the
fly by preparing string with some varibles. Using cache
variable is not possible. Because the string value is not
evaluated before accessing through lookup plugin.
So passing the cmd and where values from cmdline.

* Use start_server common task for leader election and type1 recipes.

Added changes to use common task for basic_leader_election
and healthy_raftserver_cluster_type1.yml recipes.
This start_server common task waits for complete server
bootup.

* There is no need of explicit sleep after start_server common task.

start_server.yml common task takes care of waiting for
server to bootup completely.
Removed the explicit sleeps in the recipe after
server start task now.

* Remove waiting loop for server bootup.

The start-server.yml common task already waits
for the server to bootup completely.

* Added common task start_client.yml.

Added common task to start the client and
wait for it's bootup complete.